Update server, GameService, AuthService, AuthDAO, GameDAO, MemoryGameDAO, MemoryAuthDAO, UserService

+++ server/src/main/java/service/UserService.java
@@ -3,6 +3,8 @@ package service;
 import dataaccess.DataAccessException;
 import dataaccess.MemoryAuthDAO;
 import dataaccess.MemoryUserDAO;
+import exception.ResponseException;
+import model.AuthData;
 import model.UserData;
 
 import java.util.UUID;
@@ -28,34 +30,42 @@ public class UserService {
         return UUID.randomUUID().toString();
     }
 
-    public RegisterResult register(RegisterRequest registerRequest) throws DataAccessException {
+    public void clear() {
+        memoryUserDAO.clear();
+    }
+
+    public RegisterResult register(RegisterRequest registerRequest) throws ResponseException, DataAccessException {
+        if ((registerRequest.username() == null) || (registerRequest.password() == null) ||
+                                                                                    (registerRequest.email() == null))
+            throw new ResponseException(400, "Error: bad request");
         try {
             memoryUserDAO.getUser(registerRequest.username(), registerRequest.password());
-            return new RegisterResult(null, null, "Error: already taken");
+            throw new ResponseException(403, "Error: already taken");
         } catch (DataAccessException e) {
             UserData user = memoryUserDAO.createUser(registerRequest.username(), registerRequest.password(),
-                    registerRequest.email());
+                                                                                             registerRequest.email());
             memoryAuthDAO.createAuth(generateToken(), user.username());
             return new RegisterResult(user.username(), memoryAuthDAO.getAuth(user.username()).authToken(), "OK");
         }
     }
 
-    public LoginResult login(LoginRequest loginRequest) {
+    public LoginResult login(LoginRequest loginRequest) throws ResponseException {
         try {
             UserData user = memoryUserDAO.getUser(loginRequest.username(), loginRequest.password());
-            memoryAuthDAO.createAuth(generateToken(), user.username());
-            return new LoginResult(user.username(), memoryAuthDAO.getAuth(user.username()).authToken(), "OK");
+            AuthData existingAuth = memoryAuthDAO.getAuth(user.username());
+            AuthData authData = memoryAuthDAO.createAuth(generateToken(), user.username());
+            if (existingAuth.equals(authData))
+                throw new ResponseException(401, "Error: unauthorized");
+            return new LoginResult(user.username(), authData.authToken(), "OK");
         } catch (DataAccessException e) {
-            return new LoginResult(null, null, "Error: unauthorized");
+            throw new ResponseException(401, "Error: unauthorized");
         }
     }
 
-    public LogoutResult logout(LogoutRequest logoutRequest) {
-        try {
-            memoryAuthDAO.deleteAuth(logoutRequest.authToken);
-            return new LogoutResult("OK");
-        } catch (DataAccessException e) {
-            return new LogoutResult("Error: unauthorized");
-        }
+    public LogoutResult logout(LogoutRequest logoutRequest) throws ResponseException, DataAccessException {
+        if (memoryAuthDAO.isAuthorized(logoutRequest.authToken()) == null)
+            throw new ResponseException(401, "Error: unauthorized");
+        memoryAuthDAO.deleteAuth(logoutRequest.authToken);
+        return new LogoutResult("OK");
     }
 }
